[netrw: fix window sizes]
If a file is opened from within netrw while netrw is the only window in
the current tab, the file is opened in a window of the size of
g:netrw_winsize.

Inverting g:netrw_winsize would fix this, however, if netrw is opened
while a window is already opened, the inverse happens, i.e. netrw is
opened with the size of g:netrw_winsize while the original window's size
is reduced.

To keep the netrw window to its intended size of g:netrw_winsize an
autocmd that resizes the window before leaving it, is added.