[mappings: search result navigation]
Update maps to navigate between search results and diffs.

Previously, cr and backspace mappings to jump to next/prev search result
were disabled in diff mode to use them for jumping between diffs. With
the change, diff navigation is mapped to n and N. Search navigation is
update to jump between search results using cr and backspace and their
shift-variants to also center the cursor on the new location.