[syntax: fix dts include highlighting]
issue:

'include' directives which use '<>' for the header part and are used
within a dts arch, memory or device section are not highlighted
correctly. In particular, the header part is  highlighted as part of the
respective section, i.e. could for instance be highlighted as a device
name.

reason:

'<>' doesn't match dts_string, hence the vim parser falls back to the
outside region, instead of continuing with a dts_string.

fix:

Introduce dts_header and use it for the 'nextgroup' attribute of the
pattern used to highlight 'include' directives.