[syntax: fix dts define regions]
issue:

'define' preprocessor statements are not highlighted correctly. In
particular parts of it can be highlighted for instance as dts device
names or other dts keywords.

reason:

The syntax region used to highlight 'define' statements is allowed to
contain dts_blocks.

fix:

Move the 'define' statement to a preprocessor exclusive region, which is
not allowed to contain other groups.