[vimrc: remove gtt mapping]
Remove the 'gtt' mapping, used to jump to a tag in global scope.

The mapping conflicts with the default be-complete mapping for the goto
command. Since both mappings share the first characters vim has to use a
timeout when the user types 'gt' in order to know if 'gt' or 'gtt' is
the desired sequence. Hence, every time the be-complete goto command is
desired, the user has to wait for the timeout to expire before the
command is triggered.
Since 'gtt' is not really in use, it just gets removed.