[vimrc: update make mappings]
Update the key maps to cycle between make warnings and errors.

The previous mapping for warnings partially collides with the mappings
to resize windows, e.g. 'w='. Updating the mapping avoids the timeout
that vim employs to decide if a single 'w' shall be processed or not.

For sake of consistency also update the respective make error mapping.